Fast path handling for IFIDs (#27)
* Fast path handling for IFIDs

Updating the timestamp is very cheap (same order of magnitude as enqueuing the packet for asynchronous processing), so it can safely be done on the packet-receive thread.
This helps to process IFID packets for active interfaces in time, as it and avoids the problem that interfaces time out because the IFIDs are stuck in a long packet processing queue (behind PCBs).
For inactive interfaces, the processing of the IFIDs is more work as state updates will be sent out. Also, this is not time-critical as the interface is, well, down.