Ignore ENETUNREACH, EHOSTUNREACH in border io (#3448)
When forwarding an incoming packet, ENETUNREACH can be caused by trying
to send the packet to a different address family than what the socket is
listening on (e.g. IPv6 destination address, but border routers internal
address is configured to use IPv4).
As the destination address is controlled by an end host outside the AS,
the router should not abort in this case.

Similarly, EHOSTUNREACH can be caused by a bad destination address;
from the manual:
  No valid routing table entry matches the destination address.  This
  error can be caused by an ICMP message from a remote router or for the
  local routing table.

So this error can be caused by any of:
- a misconfiguration of IP routes on the border router
- specific configuration on the border router to ensure packets are only
  forwarded to the local AS, and a destination address that does not match
- a (temporary) network failure in the local AS

As there is no clear way to distinguish misconfigurations from bad
destination addresses or transient errors, it seems best to just treat
always this as recoverable.